Refactor: Middleware
